horse shoe

+++ Rating 900/B. Drinks.cpp
@@ -9,6 +9,5 @@ int main(){
         cin>>arr[i];
         count+=arr[i];
     }
-    cout<<count/n;
-
+    cout<<(double)count/n;
 }
